Creating new users now works. Fixes UIU-291
