+++ lib/SearchAndSort/SearchAndSort.js
@@ -6,6 +6,7 @@ import PropTypes from 'prop-types';
 import Route from 'react-router-dom/Route';
 import { withRouter } from 'react-router'
 import queryString from 'query-string';
+import uuid from 'uuid';
 import { stripesShape } from '@folio/stripes-core/src/Stripes';
 import FilterGroups, { initialFilterState, onChangeFilter as commonChangeFilter } from '@folio/stripes-components/lib/FilterGroups';
 import FilterPaneSearch from '@folio/stripes-components/lib/FilterPaneSearch';
@@ -18,6 +19,7 @@ import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
 import Layer from '@folio/stripes-components/lib/Layer';
 import SRStatus from '@folio/stripes-components/lib/SRStatus';
 import transitionToParams from '@folio/stripes-components/util/transitionToParams';
+import removeQueryParam from '../../removeQueryParam';
 import UserForm from '../../UserForm'; // XXX relocate this
 import ViewUser from '../../ViewUser'; // XXX relocate this
 import contactTypes from '../../data/contactTypes';
@@ -186,6 +188,82 @@ class SearchAndSort extends React.Component {
     this.props.history.push(`/users/view/${userId}${this.props.location.search}`);
   }
 
+  onClickAddNewUser = (e) => {
+    if (e) e.preventDefault();
+    this.log('action', 'clicked "add new user"');
+    this.transitionToParams({ layer: 'create' });
+  }
+
+  onClickCloseNewUser = (e) => {
+    if (e) e.preventDefault();
+    this.log('action', 'clicked "close new user"');
+    removeQueryParam('layer', this.props.location, this.props.history);
+  }
+
+  // XXX something bad is happening here that prevents exceptions in this function from being received
+  create = (userdata) => {
+    if (userdata.personal.addresses) {
+      const addressTypes = (this.props.resources.addressTypes || {}).records || [];
+      userdata.personal.addresses = toUserAddresses(userdata.personal.addresses, addressTypes); // eslint-disable-line no-param-reassign
+    }
+    const creds = Object.assign({}, userdata.creds, { username: userdata.username });
+    const user = Object.assign({}, userdata, { id: uuid() });
+    if (user.creds) delete user.creds;
+
+    this.postUser(user)
+    .then(newUser => this.postCreds(newUser.id, creds))
+    .then(userId => this.postPerms(userId))
+    .then((userId) => {
+      this.onClickCloseNewUser();
+      this.onSelectRow(null, { id: userId });
+    });
+  }
+
+  postUser = user =>
+    fetch(`${this.props.okapi.url}/users`, {
+      method: 'POST',
+      headers: Object.assign({}, { 'X-Okapi-Tenant': this.props.okapi.tenant, 'X-Okapi-Token': this.props.okapi.token, 'Content-Type': 'application/json' }),
+      body: JSON.stringify(user),
+    }).then((userPostResponse) => {
+      if (userPostResponse.status >= 400) {
+        throw new SubmissionError('Creating new user failed');
+      } else {
+        return userPostResponse.json();
+      }
+    }).then(userJson => userJson);
+
+  postCreds = (userId, creds) => {
+    this.log('xhr', `POST credentials for new user '${userId}':`, creds);
+    const localCreds = Object.assign({}, creds, creds.password ? {} : { password: '' }, { userId });
+    return fetch(`${this.props.okapi.url}/authn/credentials`, {
+      method: 'POST',
+      headers: Object.assign({}, { 'X-Okapi-Tenant': this.props.okapi.tenant, 'X-Okapi-Token': this.props.okapi.token, 'Content-Type': 'application/json' }),
+      body: JSON.stringify(localCreds),
+    }).then((credsPostResponse) => {
+      if (credsPostResponse.status >= 400) {
+        throw new SubmissionError('Creating credentials for new user failed');
+      } else {
+        return userId;
+      }
+    });
+  }
+
+  postPerms = (userId) => {
+    const permissions = [];
+    this.log('xhr', `POST permissions for new user '${userId}':`, permissions);
+    return fetch(`${this.props.okapi.url}/perms/users`, {
+      method: 'POST',
+      headers: Object.assign({}, { 'X-Okapi-Tenant': this.props.okapi.tenant, 'X-Okapi-Token': this.props.okapi.token, 'Content-Type': 'application/json' }),
+      body: JSON.stringify({ userId, permissions }),
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new SubmissionError('Creating empty permissions for new user failed');
+      } else {
+        return userId;
+      }
+    });
+  }
+
   collapseDetails = () => {
     this.setState({
       selectedItem: {},
@@ -314,7 +392,7 @@ class SearchAndSort extends React.Component {
             addressTypes={addressTypes}
             onSubmit={(record) => { this.create(record); }}
             onCancel={this.onClickCloseNewUser}
-            okapi={this.okapi}
+            okapi={this.props.okapi}
             optionLists={{ patronGroups, contactTypes }}
           />
         </Layer>
